refactor: 테스트 메소드명 변경 (#337)
