+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/unit/post/domain/PostRepositoryTest.java
@@ -169,88 +169,6 @@ class PostRepositoryTest {
         assertThat(comments).hasSize(1);
     }
 
-    @DisplayName("Tag를 기반으로 Post를 불러온다.")
-    @ParameterizedTest
-    @MethodSource("getPostsAndTagsForSearchPostByTagName")
-    void findAllPostsByTagNamesfindAllPostsByTagNames_findAllPostsByTagNames_Success(
-        List<Post> posts,
-        User user,
-        List<Tag> tags,
-        String keyword,
-        List<PostResponseDto> expected
-    ) {
-        userRepository.save(user);
-        tags.forEach(tagRepository::save);
-        posts.forEach(postRepository::save);
-
-        PageRequest pageRequest = PageRequest.of(0, 3);
-        List<String> keywords = List.of(keyword.split(" "));
-        List<Post> savedPosts = postRepository.findAllPostsByTagNames(keywords, pageRequest);
-
-        List<PostResponseDto> actual =
-            PostDtoAssembler.assembleFrom(user, false, savedPosts);
-
-        actual.sort(comparing(PostResponseDto::getAuthorName));
-        expected.sort(comparing(PostResponseDto::getAuthorName));
-
-        assertThat(actual)
-            .usingRecursiveComparison()
-            .ignoringFields("id", "createdAt", "updatedAt")
-            .isEqualTo(expected);
-    }
-
-    public static Stream<Arguments> getPostsAndTagsForSearchPostByTagName() {
-        return Stream.of(
-            createArgument("tag1", 0),
-            createArgument("tag1 tag9", 0, 2),
-            createArgument("tag3 tag1", 0),
-            createArgument("tag1 tag4", 0, 1),
-            createArgument("tag10"),
-            createArgument("tag10 tag1", 0)
-        );
-    }
-
-    private static Arguments createArgument(String keyword, int... expectedIndex) {
-        User user = UserFactory.user("testUser");
-
-        Tag[] tags = IntStream.rangeClosed(1, 9)
-            .mapToObj(i -> new Tag(String.format("tag%d", i)))
-            .toArray(Tag[]::new);
-
-        Post[] posts = new Post[]{
-            Post.builder()
-                .content("testContent1")
-                .githubRepoUrl("https://github.com/bperhaps1")
-                .author(user)
-                .tags(tags[0], tags[1], tags[2])
-                .build(),
-            Post.builder()
-                .content("testContent2")
-                .githubRepoUrl("https://github.com/bperhaps2")
-                .author(user)
-                .tags(tags[3], tags[4], tags[5])
-                .build(),
-            Post.builder()
-                .content("testContent3")
-                .githubRepoUrl("https://github.com/bperhaps3")
-                .author(user)
-                .tags(tags[6], tags[7], tags[8])
-                .build()
-        };
-
-        List<Post> expected = Arrays.stream(expectedIndex)
-            .mapToObj(i -> posts[i])
-            .collect(toList());
-
-        return Arguments.of(
-            List.of(posts[0], posts[1], posts[2]),
-            user,
-            List.of(tags),
-            keyword,
-            PostDtoAssembler.assembleFrom(user, false, expected)
-        );
-    }
-
     private void flushAndClear() {
         testEntityManager.flush();
         testEntityManager.clear();
